+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/dto/PaymentApplicationMapper.java
@@ -1,4 +1,4 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.application.dto;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
