+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/dto/kakao/KakaoPayApproveRequest.java
@@ -2,6 +2,9 @@ package com.boeingmerryho.business.paymentservice.application.dto.kakao;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import lombok.Builder;
+
+@Builder
 public record KakaoPayApproveRequest(
 	@JsonProperty("cid") String cid,
 	@JsonProperty("tid") String tid,
