+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/application/dto/kakao/KakaoPaymentSession.java
@@ -2,6 +2,9 @@ package com.boeingmerryho.business.paymentservice.application.dto.kakao;
 
 import java.util.List;
 
+import lombok.Builder;
+
+@Builder
 public record KakaoPaymentSession(
 	String tid,
 	String cid,
