+++ payment-service/src/main/java/com/boeingmerryho/business/paymentservice/infrastructure/KakaoPaymentHelper.java
@@ -1,9 +1,10 @@
-package com.boeingmerryho.business.paymentservice.application;
+package com.boeingmerryho.business.paymentservice.infrastructure;
 
 import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
+import com.boeingmerryho.business.paymentservice.application.KakaoPaySessionService;
 import com.boeingmerryho.business.paymentservice.application.dto.kakao.KakaoPaymentSession;
 
 import lombok.RequiredArgsConstructor;
@@ -12,17 +13,18 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class KakaoPaymentHelper {
 
-	private final KakaoPaymentSessionService kakaoPaymentSessionService;
+	private final KakaoPaySessionService kakaoPaySessionService;
 
 	public void savePaymentInfo(String paymentId, KakaoPaymentSession session) {
-		kakaoPaymentSessionService.saveSession(paymentId, session);
+		kakaoPaySessionService.saveSession(paymentId, session);
 	}
 
 	public Optional<KakaoPaymentSession> getPaymentInfo(String paymentId) {
-		return kakaoPaymentSessionService.getSession(paymentId);
+		return kakaoPaySessionService.getSession(paymentId);
 	}
 
 	public void deletePaymentInfo(String paymentId) {
-		kakaoPaymentSessionService.deleteSession(paymentId);
+		kakaoPaySessionService.deleteSession(paymentId);
 	}
+
 }
